requirements.txt
aioredis
asyncio
discord.py
psutil
PyNaCl
python-dateutil
requests
youtube-dl
